Moved Locale folder
