+++ index.js
@@ -3,7 +3,7 @@ import {connectTool} from 'tc-tool';
 import path from 'path';
 
 const NAMESPACE = 'ScripturePane';
-const LOCALE_DIR = path.join(__dirname, './locale');
+const LOCALE_DIR = path.join(__dirname, './src/locale');
 
 export default {
   name: NAMESPACE,
